+++ src/gen-charts.ts
@@ -18,9 +18,9 @@ import {
 	DEF_SHAPE_SHADOW,
 	LETTERS,
 } from './core-enums'
-import { IChartOptsLib, ISlideRelChart, ShadowProps, OptsChartData, IChartTitleOpts, OptsChartGridLine } from './core-interfaces'
+import { IChartOptsLib, ISlideRelChart, ShadowProps, OptsChartData, IChartPropsTitle, OptsChartGridLine } from './core-interfaces'
 import { createColorElement, genXmlColorSelection, convertRotationDegrees, encodeXmlEntities, getMix, getUuid, valToPts } from './gen-utils'
-import * as JSZip from 'jszip'
+import JSZip from 'jszip'
 
 /**
  * Based on passed data, creates Excel Worksheet that is used as a data source for a chart.
@@ -408,7 +408,7 @@ export function createExcelWorksheet(chartObject: ISlideRelChart, zip: JSZip): P
 				zip.file('ppt/charts/' + chartObject.fileName, makeXmlCharts(chartObject))
 
 				// 3: Done
-				resolve()
+				resolve(null)
 			})
 			.catch(strErr => {
 				reject(strErr)
@@ -438,12 +438,13 @@ export function makeXmlCharts(rel: ISlideRelChart): string {
 		if (rel.opts.showTitle) {
 			strXml += genXmlTitle({
 				title: rel.opts.title || 'Chart Title',
-				fontSize: rel.opts.titleFontSize || DEF_FONT_TITLE_SIZE,
 				color: rel.opts.titleColor,
 				fontFace: rel.opts.titleFontFace,
-				rotate: rel.opts.titleRotate,
+				fontSize: rel.opts.titleFontSize || DEF_FONT_TITLE_SIZE,
 				titleAlign: rel.opts.titleAlign,
+				titleBold: rel.opts.titleBold,
 				titlePos: rel.opts.titlePos,
+				titleRotate: rel.opts.titleRotate,
 			})
 			strXml += '<c:autoTitleDeleted val="0"/>'
 		} else {
@@ -489,7 +490,7 @@ export function makeXmlCharts(rel: ISlideRelChart): string {
 			//let options: IChartOptsLib = { type: type.type, }
 			let valAxisId = options['secondaryValAxis'] ? AXIS_ID_VALUE_SECONDARY : AXIS_ID_VALUE_PRIMARY
 			let catAxisId = options['secondaryCatAxis'] ? AXIS_ID_CATEGORY_SECONDARY : AXIS_ID_CATEGORY_PRIMARY
-			usesSecondaryValAxis = usesSecondaryValAxis || options['secondaryValAxis']
+			usesSecondaryValAxis = usesSecondaryValAxis || options.secondaryValAxis
 			strXml += makeChartType(type.type, type.data, options, valAxisId, catAxisId, true)
 		})
 	} else {
@@ -499,7 +500,7 @@ export function makeXmlCharts(rel: ISlideRelChart): string {
 	// B: Axes -----------------------------------------------------------
 	if (rel.opts._type !== CHART_TYPE.PIE && rel.opts._type !== CHART_TYPE.DOUGHNUT) {
 		// Param check
-		if (rel.opts.valAxes && !usesSecondaryValAxis) {
+		if (rel.opts.valAxes && rel.opts.valAxes.length > 1 && !usesSecondaryValAxis) {
 			throw new Error('Secondary axis must be used by one of the multiple charts')
 		}
 
@@ -550,7 +551,9 @@ export function makeXmlCharts(rel: ISlideRelChart): string {
 			strXml += '	  <a:lstStyle/>'
 			strXml += '	  <a:p>'
 			strXml += '		<a:pPr rtl="0">'
-			strXml += `       <a:defRPr sz="${rel.opts.dataTableFontSize || DEF_FONT_SIZE}00" b="0" i="0" u="none" strike="noStrike" kern="1200" baseline="0">`
+			strXml += `       <a:defRPr sz="${Math.round(
+				(rel.opts.dataTableFontSize || DEF_FONT_SIZE) * 100
+			)}" b="0" i="0" u="none" strike="noStrike" kern="1200" baseline="0">`
 			strXml += '			<a:solidFill><a:schemeClr val="tx1"><a:lumMod val="65000"/><a:lumOff val="35000"/></a:schemeClr></a:solidFill>'
 			strXml += '			<a:latin typeface="+mn-lt"/>'
 			strXml += '			<a:ea typeface="+mn-ea"/>'
@@ -569,9 +572,7 @@ export function makeXmlCharts(rel: ISlideRelChart): string {
 		strXml += rel.opts.fill ? genXmlColorSelection(rel.opts.fill) : '<a:noFill/>'
 
 		// OPTION: Border
-		strXml += rel.opts.border
-			? `<a:ln w="${valToPts(rel.opts.border.pt)}" cap="flat">${genXmlColorSelection(rel.opts.border.color)}</a:ln>`
-			: '<a:ln><a:noFill/></a:ln>'
+		strXml += rel.opts.border ? `<a:ln w="${valToPts(rel.opts.border.pt)}" cap="flat">${genXmlColorSelection(rel.opts.border.color)}</a:ln>` : '<a:ln><a:noFill/></a:ln>'
 
 		// Close shapeProp/plotArea before Legend
 		strXml += '    <a:effectLst/>'
@@ -591,7 +592,7 @@ export function makeXmlCharts(rel: ISlideRelChart): string {
 				strXml += '  <a:lstStyle/>'
 				strXml += '  <a:p>'
 				strXml += '    <a:pPr>'
-				strXml += rel.opts.legendFontSize ? '<a:defRPr sz="' + Number(rel.opts.legendFontSize) * 100 + '">' : '<a:defRPr>'
+				strXml += rel.opts.legendFontSize ? '<a:defRPr sz="' + Math.round(Number(rel.opts.legendFontSize) * 100) + '">' : '<a:defRPr>'
 				if (rel.opts.legendColor) strXml += genXmlColorSelection(rel.opts.legendColor)
 				if (rel.opts.legendFontFace) strXml += '<a:latin typeface="' + rel.opts.legendFontFace + '"/>'
 				if (rel.opts.legendFontFace) strXml += '<a:cs    typeface="' + rel.opts.legendFontFace + '"/>'
@@ -706,7 +707,7 @@ function makeChartType(chartType: CHART_NAME, data: OptsChartData[], opts: IChar
 				if (seriesColor === 'transparent') {
 					strXml += '<a:noFill/>'
 				} else if (opts.chartColorsOpacity) {
-					strXml += '<a:solidFill>' + createColorElement(seriesColor, '<a:alpha val="' + opts.chartColorsOpacity + '000"/>') + '</a:solidFill>'
+					strXml += '<a:solidFill>' + createColorElement(seriesColor, `<a:alpha val="${Math.round(opts.chartColorsOpacity * 1000)}"/>`) + '</a:solidFill>'
 				} else {
 					strXml += '<a:solidFill>' + createColorElement(seriesColor) + '</a:solidFill>'
 				}
@@ -745,7 +746,7 @@ function makeChartType(chartType: CHART_NAME, data: OptsChartData[], opts: IChar
 					strXml += '      <a:bodyPr/>'
 					strXml += '      <a:lstStyle/>'
 					strXml += '      <a:p><a:pPr>'
-					strXml += '        <a:defRPr b="0" i="0" strike="noStrike" sz="' + (opts.dataLabelFontSize || DEF_FONT_SIZE) + '00" u="none">'
+					strXml += '        <a:defRPr b="0" i="0" strike="noStrike" sz="' + Math.round((opts.dataLabelFontSize || DEF_FONT_SIZE) * 100) + '" u="none">'
 					strXml += '          <a:solidFill>' + createColorElement(opts.dataLabelColor || DEF_FONT_COLOR) + '</a:solidFill>'
 					strXml += '          <a:latin typeface="' + (opts.dataLabelFontFace || 'Arial') + '"/>'
 					strXml += '        </a:defRPr>'
@@ -882,7 +883,11 @@ function makeChartType(chartType: CHART_NAME, data: OptsChartData[], opts: IChar
 				strXml += '      <a:lstStyle/>'
 				strXml += '      <a:p><a:pPr>'
 				strXml +=
-					'        <a:defRPr b="' + (opts.dataLabelFontBold ? 1 : 0) + '" i="0" strike="noStrike" sz="' + (opts.dataLabelFontSize || DEF_FONT_SIZE) + '00" u="none">'
+					'        <a:defRPr b="' +
+					(opts.dataLabelFontBold ? 1 : 0) +
+					'" i="0" strike="noStrike" sz="' +
+					Math.round((opts.dataLabelFontSize || DEF_FONT_SIZE) * 100) +
+					'" u="none">'
 				strXml += '          <a:solidFill>' + createColorElement(opts.dataLabelColor || DEF_FONT_COLOR) + '</a:solidFill>'
 				strXml += '          <a:latin typeface="' + (opts.dataLabelFontFace || 'Arial') + '"/>'
 				strXml += '        </a:defRPr>'
@@ -961,7 +966,7 @@ function makeChartType(chartType: CHART_NAME, data: OptsChartData[], opts: IChar
 					if (tmpSerColor === 'transparent') {
 						strXml += '<a:noFill/>'
 					} else if (opts.chartColorsOpacity) {
-						strXml += '<a:solidFill>' + createColorElement(tmpSerColor, '<a:alpha val="' + opts.chartColorsOpacity + '000"/>') + '</a:solidFill>'
+						strXml += '<a:solidFill>' + createColorElement(tmpSerColor, '<a:alpha val="' + Math.round(opts.chartColorsOpacity * 1000) + '"/>') + '</a:solidFill>'
 					} else {
 						strXml += '<a:solidFill>' + createColorElement(tmpSerColor) + '</a:solidFill>'
 					}
@@ -1196,7 +1201,7 @@ function makeChartType(chartType: CHART_NAME, data: OptsChartData[], opts: IChar
 				strXml += '      <a:bodyPr/>'
 				strXml += '      <a:lstStyle/>'
 				strXml += '      <a:p><a:pPr>'
-				strXml += '        <a:defRPr b="0" i="0" strike="noStrike" sz="' + (opts.dataLabelFontSize || DEF_FONT_SIZE) + '00" u="none">'
+				strXml += '        <a:defRPr b="0" i="0" strike="noStrike" sz="' + Math.round((opts.dataLabelFontSize || DEF_FONT_SIZE) * 100) + '" u="none">'
 				strXml += '          <a:solidFill>' + createColorElement(opts.dataLabelColor || DEF_FONT_COLOR) + '</a:solidFill>'
 				strXml += '          <a:latin typeface="' + (opts.dataLabelFontFace || 'Arial') + '"/>'
 				strXml += '        </a:defRPr>'
@@ -1261,7 +1266,7 @@ function makeChartType(chartType: CHART_NAME, data: OptsChartData[], opts: IChar
 					if (tmpSerColor === 'transparent') {
 						strXml += '<a:noFill/>'
 					} else if (opts.chartColorsOpacity) {
-						strXml += '<a:solidFill>' + createColorElement(tmpSerColor, '<a:alpha val="' + opts.chartColorsOpacity + '000"/>') + '</a:solidFill>'
+						strXml += '<a:solidFill>' + createColorElement(tmpSerColor, '<a:alpha val="' + Math.round(opts.chartColorsOpacity * 1000) + '"/>') + '</a:solidFill>'
 					} else {
 						strXml += '<a:solidFill>' + createColorElement(tmpSerColor) + '</a:solidFill>'
 					}
@@ -1350,7 +1355,7 @@ function makeChartType(chartType: CHART_NAME, data: OptsChartData[], opts: IChar
 				strXml += '      <a:bodyPr/>'
 				strXml += '      <a:lstStyle/>'
 				strXml += '      <a:p><a:pPr>'
-				strXml += '        <a:defRPr b="0" i="0" strike="noStrike" sz="' + (opts.dataLabelFontSize || DEF_FONT_SIZE) + '00" u="none">'
+				strXml += '        <a:defRPr b="0" i="0" strike="noStrike" sz="' + Math.round((opts.dataLabelFontSize || DEF_FONT_SIZE) * 100) + '" u="none">'
 				strXml += '          <a:solidFill>' + createColorElement(opts.dataLabelColor || DEF_FONT_COLOR) + '</a:solidFill>'
 				strXml += '          <a:latin typeface="' + (opts.dataLabelFontFace || 'Arial') + '"/>'
 				strXml += '        </a:defRPr>'
@@ -1450,7 +1455,9 @@ function makeChartType(chartType: CHART_NAME, data: OptsChartData[], opts: IChar
 				strXml += '  <c:spPr/><c:txPr>'
 				strXml += '   <a:bodyPr/><a:lstStyle/>'
 				strXml += '   <a:p><a:pPr>'
-				strXml += `   <a:defRPr sz="${opts.dataLabelFontSize || DEF_FONT_SIZE}00" b="${opts.dataLabelFontBold ? 1 : 0}" i="0" u="none" strike="noStrike">`
+				strXml += `   <a:defRPr sz="${Math.round((opts.dataLabelFontSize || DEF_FONT_SIZE) * 100)}" b="${
+					opts.dataLabelFontBold ? 1 : 0
+				}" i="0" u="none" strike="noStrike">`
 				strXml += '    <a:solidFill>' + createColorElement(opts.dataLabelColor || DEF_FONT_COLOR) + '</a:solidFill>'
 				strXml += `    <a:latin typeface="${opts.dataLabelFontFace || 'Arial'}"/>`
 				strXml += '   </a:defRPr>'
@@ -1515,7 +1522,7 @@ function makeChartType(chartType: CHART_NAME, data: OptsChartData[], opts: IChar
 
 			// 4: Close "SERIES"
 			strXml += '  </c:ser>'
-			strXml += '  <c:firstSliceAng val="0"/>'
+			strXml += `  <c:firstSliceAng val="${opts.firstSliceAng ? Math.round(opts.firstSliceAng) : 0}"/>`
 			if (chartType === CHART_TYPE.DOUGHNUT) strXml += '  <c:holeSize val="' + (opts.holeSize || 50) + '"/>'
 			strXml += '</c:' + chartType + 'Chart>'
 
@@ -1561,7 +1568,7 @@ function makeCatAxis(opts: IChartOptsLib, axisId: string, valAxisId: string): st
 			color: opts.catAxisTitleColor,
 			fontFace: opts.catAxisTitleFontFace,
 			fontSize: opts.catAxisTitleFontSize,
-			rotate: opts.catAxisTitleRotate,
+			titleRotate: opts.catAxisTitleRotate,
 			title: opts.catAxisTitle || 'Axis Title',
 		})
 	}
@@ -1578,7 +1585,7 @@ function makeCatAxis(opts: IChartOptsLib, axisId: string, valAxisId: string): st
 	} else {
 		strXml += '  <c:majorTickMark val="' + (opts.catAxisMajorTickMark || 'out') + '"/>'
 		strXml += '  <c:minorTickMark val="' + (opts.catAxisMajorTickMark || 'none') + '"/>'
-		strXml += '  <c:tickLblPos val="' + (opts.catAxisLabelPos || opts.barDir === 'col' ? 'low' : 'nextTo') + '"/>'
+		strXml += '  <c:tickLblPos val="' + (opts.catAxisLabelPos || (opts.barDir === 'col' ? 'low' : 'nextTo')) + '"/>'
 	}
 	strXml += '  <c:spPr>'
 	strXml += '    <a:ln w="' + (opts.catAxisLineSize !== undefined ? valToPts(opts.catAxisLineSize) : 12700) + '" cap="flat">'
@@ -1592,7 +1599,12 @@ function makeCatAxis(opts: IChartOptsLib, axisId: string, valAxisId: string): st
 	strXml += '    <a:lstStyle/>'
 	strXml += '    <a:p>'
 	strXml += '    <a:pPr>'
-	strXml += '    <a:defRPr sz="' + (opts.catAxisLabelFontSize || DEF_FONT_SIZE) + '00" b="' + (opts.catAxisLabelFontBold ? 1 : 0) + '" i="0" u="none" strike="noStrike">'
+	strXml +=
+		'    <a:defRPr sz="' +
+		Math.round((opts.catAxisLabelFontSize || DEF_FONT_SIZE) * 100) +
+		'" b="' +
+		(opts.catAxisLabelFontBold ? 1 : 0) +
+		'" i="0" u="none" strike="noStrike">'
 	strXml += '      <a:solidFill><a:srgbClr val="' + (opts.catAxisLabelColor || DEF_FONT_COLOR) + '"/></a:solidFill>'
 	strXml += '      <a:latin typeface="' + (opts.catAxisLabelFontFace || 'Arial') + '"/>'
 	strXml += '   </a:defRPr>'
@@ -1663,7 +1675,7 @@ function makeValAxis(opts: IChartOptsLib, valAxisId: string): string {
 			color: opts.valAxisTitleColor,
 			fontFace: opts.valAxisTitleFontFace,
 			fontSize: opts.valAxisTitleFontSize,
-			rotate: opts.valAxisTitleRotate,
+			titleRotate: opts.valAxisTitleRotate,
 			title: opts.valAxisTitle || 'Axis Title',
 		})
 	}
@@ -1675,7 +1687,7 @@ function makeValAxis(opts: IChartOptsLib, valAxisId: string): string {
 	} else {
 		strXml += ' <c:majorTickMark val="' + (opts.valAxisMajorTickMark || 'out') + '"/>'
 		strXml += ' <c:minorTickMark val="' + (opts.valAxisMinorTickMark || 'none') + '"/>'
-		strXml += ' <c:tickLblPos val="' + (opts.valAxisLabelPos || opts.barDir === 'col' ? 'nextTo' : 'low') + '"/>'
+		strXml += ' <c:tickLblPos val="' + (opts.valAxisLabelPos || (opts.barDir === 'col' ? 'nextTo' : 'low')) + '"/>'
 	}
 	strXml += ' <c:spPr>'
 	strXml += '   <a:ln w="' + (opts.valAxisLineSize !== undefined ? valToPts(opts.valAxisLineSize) : 12700) + '" cap="flat">'
@@ -1689,7 +1701,12 @@ function makeValAxis(opts: IChartOptsLib, valAxisId: string): string {
 	strXml += '  <a:lstStyle/>'
 	strXml += '  <a:p>'
 	strXml += '    <a:pPr>'
-	strXml += '      <a:defRPr sz="' + (opts.valAxisLabelFontSize || DEF_FONT_SIZE) + '00" b="' + (opts.valAxisLabelFontBold ? 1 : 0) + '" i="0" u="none" strike="noStrike">'
+	strXml +=
+		'      <a:defRPr sz="' +
+		Math.round((opts.valAxisLabelFontSize || DEF_FONT_SIZE) * 100) +
+		'" b="' +
+		(opts.valAxisLabelFontBold ? 1 : 0) +
+		'" i="0" u="none" strike="noStrike">'
 	strXml += '        <a:solidFill><a:srgbClr val="' + (opts.valAxisLabelColor || DEF_FONT_COLOR) + '"/></a:solidFill>'
 	strXml += '        <a:latin typeface="' + (opts.valAxisLabelFontFace || 'Arial') + '"/>'
 	strXml += '      </a:defRPr>'
@@ -1737,7 +1754,7 @@ function makeSerAxis(opts: IChartOptsLib, axisId: string, valAxisId: string): st
 			color: opts.serAxisTitleColor,
 			fontFace: opts.serAxisTitleFontFace,
 			fontSize: opts.serAxisTitleFontSize,
-			rotate: opts.serAxisTitleRotate,
+			titleRotate: opts.serAxisTitleRotate,
 			title: opts.serAxisTitle || 'Axis Title',
 		})
 	}
@@ -1757,7 +1774,7 @@ function makeSerAxis(opts: IChartOptsLib, axisId: string, valAxisId: string): st
 	strXml += '    <a:lstStyle/>'
 	strXml += '    <a:p>'
 	strXml += '    <a:pPr>'
-	strXml += '    <a:defRPr sz="' + (opts.serAxisLabelFontSize || DEF_FONT_SIZE) + '00" b="0" i="0" u="none" strike="noStrike">'
+	strXml += `    <a:defRPr sz="${Math.round((opts.serAxisLabelFontSize || DEF_FONT_SIZE) * 100)}" b="0" i="0" u="none" strike="noStrike">`
 	strXml += '      <a:solidFill><a:srgbClr val="' + (opts.serAxisLabelColor || DEF_FONT_COLOR) + '"/></a:solidFill>'
 	strXml += '      <a:latin typeface="' + (opts.serAxisLabelFontFace || 'Arial') + '"/>'
 	strXml += '   </a:defRPr>'
@@ -1793,13 +1810,14 @@ function makeSerAxis(opts: IChartOptsLib, axisId: string, valAxisId: string): st
 
 /**
  * Create char title elements
- * @param {IChartTitleOpts} opts - options
+ * @param {IChartPropsTitle} opts - options
  * @return {string} XML `<c:title>`
  */
-function genXmlTitle(opts: IChartTitleOpts): string {
+function genXmlTitle(opts: IChartPropsTitle): string {
 	let align = opts.titleAlign === 'left' || opts.titleAlign === 'right' ? `<a:pPr algn="${opts.titleAlign.substring(0, 1)}">` : `<a:pPr>`
-	let rotate = opts.rotate ? `<a:bodyPr rot="${convertRotationDegrees(opts.rotate)}"/>` : `<a:bodyPr/>` // don't specify rotation to get default (ex. vertical for cat axis)
-	let sizeAttr = opts.fontSize ? 'sz="' + Math.round(opts.fontSize) + '00"' : '' // only set the font size if specified.  Powerpoint will handle the default size
+	let rotate = opts.titleRotate ? `<a:bodyPr rot="${convertRotationDegrees(opts.titleRotate)}"/>` : `<a:bodyPr/>` // don't specify rotation to get default (ex. vertical for cat axis)
+	let sizeAttr = opts.fontSize ? 'sz="' + Math.round(opts.fontSize * 100) + '"' : '' // only set the font size if specified.  Powerpoint will handle the default size
+	let titleBold = opts.titleBold === true ? 1 : 0
 	let layout =
 		opts.titlePos && opts.titlePos.x && opts.titlePos.y
 			? `<c:layout><c:manualLayout><c:xMode val="edge"/><c:yMode val="edge"/><c:x val="${opts.titlePos.x}"/><c:y val="${opts.titlePos.y}"/></c:manualLayout></c:layout>`
@@ -1812,13 +1830,13 @@ function genXmlTitle(opts: IChartTitleOpts): string {
 	      <a:lstStyle/>
 	      <a:p>
 	        ${align}
-	        <a:defRPr ${sizeAttr} b="0" i="0" u="none" strike="noStrike">
+	        <a:defRPr ${sizeAttr} b="${titleBold}" i="0" u="none" strike="noStrike">
 	          <a:solidFill><a:srgbClr val="${opts.color || DEF_FONT_COLOR}"/></a:solidFill>
 	          <a:latin typeface="${opts.fontFace || 'Arial'}"/>
 	        </a:defRPr>
 	      </a:pPr>
 	      <a:r>
-	        <a:rPr ${sizeAttr} b="0" i="0" u="none" strike="noStrike">
+	        <a:rPr ${sizeAttr} b="${titleBold}" i="0" u="none" strike="noStrike">
 	          <a:solidFill><a:srgbClr val="${opts.color || DEF_FONT_COLOR}"/></a:solidFill>
 	          <a:latin typeface="${opts.fontFace || 'Arial'}"/>
 	        </a:rPr>
@@ -1871,9 +1889,9 @@ function createShadowElement(options: ShadowProps, defaults: object): string {
 		type = opts['type'] || 'outer',
 		blur = valToPts(opts['blur']),
 		offset = valToPts(opts['offset']),
-		angle = opts['angle'] * 60000,
+		angle = Math.round(opts['angle'] * 60000),
 		color = opts['color'],
-		opacity = opts['opacity'] * 100000,
+		opacity = Math.round(opts['opacity'] * 100000),
 		rotateWithShape = opts['rotateWithShape'] ? 1 : 0
 
 	strXml += '<a:' + type + 'Shdw sx="100000" sy="100000" kx="0" ky="0"  algn="bl" blurRad="' + blur + '" '